pyproject.toml
[tool.poetry]
name = "pyosrs"
version = "0.0.0"
description = "A Python library for Oldschool RuneScape."
authors = ["Phong Tran"]
repository = "https://github.com/phongse/pyosrs"
license = "LGPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
httpx = "^0.24.0"
pydantic = "^1.10.7"
python = "^3.10"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
coverage = "^7.2.3"
factory-boy = "^3.2.1"
isort = "^5.12.0"
mypy = "^1.2.0"
pre-commit = "^3.2.2"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.10.0"
respx = "^0.20.1"

[tool.isort]
profile = "black"

[tool.mypy]
exclude = [
    "testing/",
    "tests/",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
